Add shared memories
This change adds the ability to use shared memories in Wasmtime when the
[threads proposal] is enabled. Shared memories are annotated as `shared`
in the WebAssembly syntax, e.g., `(memory 1 1 shared)`, and are
protected from concurrent access during `memory.size` and `memory.grow`.

In order to implement this in Wasmtime, there are two main cases to
cover:
 - a program may simply create a shared memory and possibly export it;
   this means that Wasmtime itself must be able to create shared
   memories
 - a user may create a shared memory externally and pass it in as an
   import during instantiation; this is the case when the program
   contains code like `(import "env" "memory" (memory 1 1
   shared))`--this case is handled by a new Wasmtime API
   type--`SharedMemory`

Because of the first case, this change allows any of the current
memory-creation mechanisms to work as-is. Wasmtime can still create
either static or dynamic memories in either on-demand or pooling modes,
and any of these memories can be considered shared. When shared, the
`Memory` runtime container will lock appropriately during `memory.size`
and `memory.grow` operations; since all memories use this container, it
is an ideal place for implementing the locking once and once only.

The second case is covered by the new `SharedMemory` structure. It uses
the same `Mmap` allocation under the hood as non-shared memories, but
allows the user to perform the allocation externally to Wasmtime and
share the memory across threads (via `Arc`). The pointer address in
`SharedMemory` is carefully wired through the memory plans by means of
`MemoryStyle::External`--this requires some careful thinking to avoid
unsafety. Because this path contains raw pointer values, it is expected
that `MemoryStyle::External` will not be serializable and therefore the
`SharedModule` use case is not compatible with pre-compiled WebAssembly
modules (`*.cwasm`). The primary usage for `SharedMemory` is for
satisfying module imports in a Wasmtime embedding.

[threads proposal]: https://github.com/WebAssembly/threads/blob/master/proposals/threads/Overview.md