wasi-nn: refactor `tf.rs`
This change refactors the implementation in #3977 to:
- avoid saving state unnecessarily across `wasi-nn` calls
- remove unused code
- refactor several bits and pieces
- add significantly more documentation

There are missing pieces, however: the "pass a model directory" paradigm
should be documented and the custom parsing of `signature,*` and `tag,*`
key-value pairs can be replaced by static parameters to the `load` call.
I wonder if these parameters could be optional and if not present a
sane default could be used by the implementation (?).

Another issue that bears more thought is the use of `Arc`s throughout.
It is unclear whether these are actually necessary or were originally
used for convenience.