fix: do not panic when thread exits with `wasi::proc_exit`
Threads that exit via `pthread_exit` use the `wasi::proc_exit` call to
inform the engine that nothing is amiss but that the thread wants to
stop. The `wasi-threads` implementation (host side) receives this as a
trap. This change ensures that when we receive a trap with `I32Exit(0)`
we treat this as a normal event and not a failure.