fix: wasi-threads does not need to store the host `T`
Since the host structure can be retrieved from the `Caller`, the
wasi-threads does not need to hold a reference to it in its context
state. Instead, it can retrieve it and clone it for the new thread (the
host is an `Arc<Host>` here). Also, wasi-threads can be `&self` instead
of `&mut self`.