fix: switch preview1 to use `&self` signatures
This change is a large change on top of the original PR to change all of
the signatures of the preview1 implementation to use `&self` instead of
`&mut self`. This makes it possible to access, and use, a `WasiCtx`
from an `Arc<Host>` which will only give immutable access to its
contents. This change also wraps the source of randomness in `WasiCtx`
with a `Mutex` to enable this `&self` signature.

This change is, however, incomplete:
- TODO: the preview0 implementation also needs a similar treatment
- TODO: we still need to modify all of the uses of slices in preview1
  and elsewhere to either copy the items if the slice is from a shared
  memory or use a `GuestSlice` otherwise