wiggle: copy guest slices back to shared memory
This change upgrades `UnsafeGuestSlice` in Wiggle to expose more
functionality to be able to use `std::ptr::copy` for writing bytes into
Wasm shared memory. With this in place, it is now possible to implement
the `preview1` "read" functions for shared memory. Previously, these
would panic if attempting to copy to a shared memory. This change
removes the panic and introduces some (rather complex) logic for
handling both the shared and non-shared cases:
- if reading into a Wasm non-shared memory, Wiggle guarantees that no
  other guest pointers will touch the memory region and, in the absence
  of concurrency, `preview1` can write directly to this memory
- if reading into a Wasm shared memory, the memory region can be
  concurrently modified. At @alexcrichton's request re: Rust safety,
  this change copies all of the bytes into an intermediate buffer before
  using `std::ptr::copy` to move them into Wasm memory.