WIP: use `Arc<Self>` in `WasiFile` and associated traits
This is an attempt to apply @alexcrichton's comment
[here](https://github.com/bytecodealliance/wasmtime/pull/5326#discussion_r1037500040)
which proposed using `self: &Arc<Self>` in the `WasiFile` trait. This
change, made as `self: Arc<Self>` here due to Rust method receiver
limitations, affects many locations. E.g., `WasiDir` also needs to start
using `Arc<Self>`, etc. This results in a slew of errors at all the use
locations. This commit fixes up many of these but is marked "WIP" since
not all are fixed; the latest challenge is deciding what the signature
of `WasiFile::pollable` should be when the switch to `Arc<Self>` is
made.