Add the MPK tooling--e.g., `Pkey`
In order to use MPK on an x86_64 Linux system, we need access to the
underlying `pkey_*` system calls (`sys`), control of the x86 PKRU
register (`pkru`), and a way of determining if MPK is even supported
(`is_supported`). These various parts are wrapped by the `Pkey`
abstraction, the public API for accessing all of this functionality.