wasi-nn: refactor to allow cloning graphs from a registry (WIP)
This adds some wrapper types, `Graph` and `ExecutionContext`, that are
used throughout the crate instead of `Box<dyn Backend...>`. The
interesting part of this is that we need to be able to "clone" graphs
from the registry without removing them (e.g., perhaps the registry is
reused among several instances). But the `BackendGraph` trait cannot
super-trait `Clone` due to `Sized` [constraints]. Here we choose to
implement a custom `clone`; another alternative would be to use the
`dyn-clone` trait.

The one remaining TODO is to figure out how to correctly clone OpenVINO
`ExecutableNetwork`s. Currently we use a `Arc::get_mut` but do not
handle the error case--perhaps this must be a mutex? It's unclear
whether `ExecutableNetwork::create_infer_request` must really use `&mut
self`, but better safe than sorry.