wasi-nn: add integration test
This change adds an integration test to the `wasmtime-wasi-nn` crate to
run the image classification example directly using `cargo test`.
Because not all environments will have OpenVINO installed, this test
checks if it is present and skips the test if it is not; if the `CI` or
`FORCE_WASINN_TEST_CHECK` environment variables are set, it fails
instead. The test pre-checks also retrieve the necessary MobileNet model
and tensor artifacts.