x64: refactor further, using `AssemblerOutputs`
This change takes the efforts in [#10276] a step further by
incorporating all the feedback gathered during review. The major change
is to shift towards the use of a new `AssemblerOutputs` type which is
returned by the new `x64_*_raw` ISLE constructor. This then forced some
refactoring, primarily getting rid of `IsleConstructorRaw`. One doubt:
while returning `AssemblerOutputs` obviates the need for the `enum`
variants (we always return the same type), maybe we should end up
creating more "generator" structs like `IsleConstructorRaw` in the
future (?). Another refactoring moved all the generation-related code
out of `dsl` and into the `generate` module; this should make it easier
to migrate this to the `cranelift-codegen-meta` crate later.

[#10276]: https://github.com/bytecodealliance/wasmtime/pull/10276