asm: model partial register writes
In [discussion] about how to define move instructions, we concluded that
we want to define instructions correctly in terms of their written-to
bits. This change adds `Operand::partial`, a field that is to be set
when a write-only, register-using operand is not written to completely;
in other words, some register bits are undefined by the current
instruction and would have been set at some earlier time. This allows us
to use `w(<location>)` to define instructions _and_ still determine
later in Cranelift if that operand's previous bits an incoming `Gpr` or
`Xmm` to track the unchanged bits.

This change sets these bits automatically for 8-bit and 16-bit scalar
instructions. For XMM-using instructions we will need to audit the
current list of defined instructions _and_ carefully model partial bits
in the future.

This also refactors some of the validation logic.

[discussion]: https://github.com/bytecodealliance/wasmtime/pull/10871#discussion_r2116237303