Extend x64 registers from 16 to 32
The upcoming APX feature extends the number of GPRs in x64 from 16 to
32. This is now possible in either APX encoding form: the REX2 prefix or
the EVEX prefix. For some time (since EVEX encodings became available),
it has also been possible to use `xmm0 - xmm31`; this seems like a good
time to extend both in Cranelift.

This change allows the Cranelift ABI to understand extended GPRs (`r16 -
r31`) and extended XMMs (`xmm16 - xmm31`). As noted in the APX
[documentation], all extended GPRs are caller-saved (volatile) in ABIs.
This change does not yet touch the rest of Cranelift, which still needs
a restriction on which instructions can access the first 16 registers
and which instructions can access all registers.

[documentation]: https://www.intel.com/content/www/us/en/developer/articles/technical/advanced-performance-extensions-apx.html