Add APX feature to ISA settings
This change adds the APX feature to `x64_settings::Flags`--`use_apx` and
`has_apx`. We do not yet update any of the presets to contain it; that
can come when the hardware is released (plus, the presets likely need
some updates for recent years). Additionally, this change clarifies the
CPUID syntax of the comments that describe which bits are checked.